Move position module to top level
