--- lib/lang/src/debug/mod.rs
@@ -1,3 +1,0 @@
-mod position;
-
-pub use position::*;
