+++ lib/lang/src/types/boolean.rs
@@ -1,4 +1,4 @@
-use crate::debug::Position;
+use crate::position::Position;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Deserialize, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
