+++ lib/lang/src/types/record.rs
@@ -1,4 +1,4 @@
-use crate::debug::Position;
+use crate::position::Position;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, PartialOrd, Ord, Serialize)]
